+++ postinstall.sh
@@ -1,0 +1,8 @@
+cp node_modules/bootstrap/dist/css/bootstrap.min.css css/bootstrap.min.css
+cp -r node_modules/bootstrap/dist/fonts .
+cp node_modules/hotbox-uex/hotbox.css css/hotbox.css
+cp node_modules/codemirror/lib/codemirror.css css/codemirror.css
+cp node_modules/kityminder-core/dist/kityminder.core.css css/kityminder.core.css
+cp node_modules/dist/kityminder.editor.min.css css/kityminder.editor.min.css
+cp -r node_modules/dist/images css/
+