+++ CHANGELOG.md
@@ -19,3 +19,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add workflow for packaging
 - Add changelog
 - Update dependencies
+- Unbundle vendor dependencies